split up into multiple files again, kinda wacky but i think it works?
